added new forms for creating a Study Area out of a set of parcels; and then a scenario for that Study Area. #31.
