+++ frontend/src/edit.jsx
@@ -1,29 +1,14 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 
 import {
   FocusStyleManager,
-  HTMLSelect,
-  Radio,
-  RadioGroup,
   Tab,
   Tabs,
 } from '@blueprintjs/core';
 
-import {
-  doWallpaper,
-  makeScenario,
-  getJobResults,
-  getJobStatus,
-  getPatterns,
-  createPattern,
-  convertToSingleLULC,
-} from './requests';
-import useInterval from './hooks/useInterval';
 import ScenarioTable from './scenarioTable';
 import PortfolioTable from './portfolioTable';
 import ParcelTable from './parcelTable';
-import landuseCodes from './landuseCodes';
-import WallpaperingMenu from './wallpaperingMenu';
 
 FocusStyleManager.onlyShowFocusOnTabs();
 
@@ -39,67 +24,7 @@ export default function EditMenu(props) {
   } = props;
 
   const [activeTab, setActiveTab] = useState('create');
-  const [scenarioName, setScenarioName] = useState('');
-  const [scenarioID, setScenarioID] = useState(null);
-  const [selectedPattern, setSelectedPattern] = useState(null);
-  const [patterns, setPatterns] = useState([]);
-  const [jobID, setJobID] = useState(null);
-  const [newPatternName, setNewPatternName] = useState('New Pattern 1');
-  const [singleLULC, setSingleLULC] = useState('');
-  const [conversionOption, setConversionOption] = useState('paint');
   const [parcelSet, setParcelSet] = useState({});
-
-  // On first render, get the list of available patterns
-  useEffect(async () => {
-    setPatterns(await getPatterns() || []);
-  }, []);
-
-  useInterval(async () => {
-    console.log('checking status for job', jobID);
-    const status = await getJobStatus(jobID);
-    if (status === 'success') {
-      const results = await getJobResults(jobID, scenarioID);
-      console.log(results);
-      setParcelTable(results);
-      setJobID(null);
-    }
-  }, (jobID && scenarioID) ? 1000 : null);
-
-  async function handleSubmitNew(event) {
-    event.preventDefault();
-    if (!scenarioName) {
-      alert('no scenario was selected');
-      return;
-    }
-    if (!parcelSet) {
-      alert('no parcel was selected; no changes to make');
-      return;
-    }
-    let currentScenarioID = scenarioID;
-    if (savedScenarios.every((scen) => scen.name !== scenarioName)) {
-      currentScenarioID = await makeScenario(sessionID, scenarioName, 'description');
-      setScenarioID(currentScenarioID);
-    }
-    let jid;
-    if (conversionOption === 'wallpaper' && selectedPattern) {
-      jid = await doWallpaper(undefined, selectedPattern.label, currentScenarioID);
-    }
-    if (conversionOption === 'paint' && singleLULC) {
-      jid = await convertToSingleLULC(undefined, singleLULC, currentScenarioID);
-    }
-    setJobID(jid);
-    refreshSavedScenarios();
-  }
-
-  const handleSamplePattern = async (event) => {
-    event.preventDefault();
-    await createPattern(patternSampleWKT, newPatternName, sessionID);
-    setPatterns(await getPatterns());
-    setSelectedPattern(
-      patterns.filter((pattern) => pattern.label === newPatternName)
-    );
-    togglePatternSamplingMode();
-  };
 
   const addParcel = async (parcel) => {
     setParcelSet((prev) => {
@@ -129,79 +54,18 @@ export default function EditMenu(props) {
           panel={(
             <div>
               <ScenarioTable
+                sessionID={sessionID}
                 parcelSet={parcelSet}
+                removeParcel={removeParcel}
+                patternSamplingMode={patternSamplingMode}
+                togglePatternSamplingMode={togglePatternSamplingMode}
+                refreshSavedScenarios={refreshSavedScenarios}
               />
               <ParcelTable
+                sessionID={sessionID}
                 parcel={selectedParcel}
                 addParcel={addParcel}
-                sessionID={sessionID}
               />
-              <br />
-              {
-                (Object.keys(parcelSet).length)
-                  ? (
-                    <form onSubmit={handleSubmitNew}>
-                      <RadioGroup
-                        className="sidebar-subheading"
-                        inline
-                        label="Modify the landuse of this parcel by:"
-                        onChange={(event) => setConversionOption(event.target.value)}
-                        selectedValue={conversionOption}
-                      >
-                        <Radio key="wallpaper" value="wallpaper" label="wallpaper" />
-                        <Radio key="paint" value="paint" label="paint" />
-                      </RadioGroup>
-                      <div className="conversion-panel">
-                        {
-                          (conversionOption === 'paint')
-                            ? (
-                              <HTMLSelect
-                                onChange={(event) => setSingleLULC(event.target.value)}
-                              >
-                                {Object.entries(landuseCodes)
-                                  .map(([code, data]) => <option key={code} value={code}>{data.name}</option>)}
-                              </HTMLSelect>
-                            )
-                            : (
-                              <WallpaperingMenu
-                                newPatternName={newPatternName}
-                                setNewPatternName={setNewPatternName}
-                                selectedPattern={selectedPattern}
-                                setSelectedPattern={setSelectedPattern}
-                                patternSamplingMode={patternSamplingMode}
-                                togglePatternSamplingMode={togglePatternSamplingMode}
-                                handleSamplePattern={handleSamplePattern}
-                                patterns={patterns}
-                              />
-                            )
-                        }
-                      </div>
-                      <p className="sidebar-subheading">
-                        Add this modification to a scenario
-                      </p>
-                      <datalist id="scenariolist">
-                        {Object.values(savedScenarios).map(
-                          (scenario) => (
-                            <option
-                              key={scenario.scenario_id}
-                              value={scenario.name} />
-                            ),
-                          )
-                        }
-                      </datalist>
-                      <input
-                        type="search"
-                        id="scenarioName"
-                        list="scenariolist"
-                        value={scenarioName}
-                        autoComplete="off"
-                        onChange={(event) => setScenarioName(event.currentTarget.value)}
-                      />
-                      <button type="submit">Submit</button>
-                    </form>
-                  )
-                  : <div />
-              }
             </div>
           )}
         />
